[fix] Added method and code refactoring
